requirements.txt
# Google APIs for Forms/Drive/Sheets
google-api-python-client>=2.137.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
pydantic-settings==2.10.1

# Database
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pgvector==0.2.4

# Document Processing
PyPDF2==3.0.1
openpyxl==3.1.2
pandas==2.2.2
python-multipart==0.0.9

# AI/LLM Integration
ollama==0.4.2
openai==1.59.5
langchain-openai==0.2.8
langchain-ollama==0.2.0
langchain-core>=0.3.17
langchain-community==0.3.7

# Optional: For advanced OCR (uncomment if needed)
# pytesseract==0.3.10
# opencv-python==4.8.1.78
# spacy==3.7.2
